Fix crash on Windows due to incorrect AiMesh struct size
mTextureCoordsNames in Assimp is a pointer to an array of pointers, not
an inline struct of 8 pointers. This resulted in the AiMesh struct using
64 bytes for this field instead of 8 bytes (on x64 Windows), causing
crashes when loading some models.

Oddly enough this posed no problems on Linux x86-64, perhaps it is less
pedantic when reading out of bounds in this scenario.

See also https://github.com/assimp/assimp/blob/v5.2.3/include/assimp/mesh.h#L754